Incremental patches to 20060909 i386_x64 support. Pointer arithmatic fixes + makefile fixes Fix to nanoEmacs window version crash. Fix to the text mode wrap margin offset.
git-svn-id: file:///media/bill/diskAshur2/backup/svn/microemacs/releases/me_20060909@1862 a259ec66-be10-11ea-b557-b34f8aa0b108

+++ src/bind.c
@@ -369,7 +369,7 @@ setCharMask(int f, int n)
     ss = buf1 ;
     while((c1 = *ss++) != '\0')
         if((dd = (meUByte *) strchr((char *) charMaskFlags,c1)) != NULL)
-            flags |= 1 << (((int) dd)- ((int) charMaskFlags)) ;
+            flags |= 1 << ((int)(dd - charMaskFlags)) ;
     mask2 = (flags & 0x00ff) ;
     mask1 = (flags & 0x0f00) >> 4 ;
